requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Login==0.6.*
Flask-WTF==1.2.*
Flask-Migrate==4.0.*
Flask-Limiter==3.5.*
Flask-Talisman==1.1.*
argon2-cffi==23.*
python-dotenv==1.0.*
Werkzeug==3.1.*
email-validator==2.*
fpdf2==2.*
